requirements.txt
google-genai
google-api-core
newspaper4k
notion-client
python-dotenv
googlenewsdecoder
feedparser